requirements.txt
requests==2.31.0
python-dotenv==1.0.0
PyGithub==2.1.1
textblob==0.17.1
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4
pytest==8.3.4
flask==3.0.0
flask-cors==4.0.0
